infrastructure/containers/3dgs_legacy/requirements.txt
# AWS SDK
boto3>=1.26.0

# Core dependencies
numpy>=1.24.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.64.0

# 3D Gaussian Splatting
gsplat>=0.1.0
plyfile>=0.7.4

# System utilities for multiprocessing
psutil>=5.9.0 